requirements.txt
numpy==1.19.5
Pillow>=8.2.0
tifffile>=2020.9.3